tests: move to agg_tests.py
